Modules to create persons and books added
