+++ app.rb
@@ -1,10 +1,13 @@
 require_relative 'classes/book'
-require_relative 'classes/person'
-require_relative 'classes/teacher'
-require_relative 'classes/student'
+# require_relative 'classes/person'
+# require_relative 'classes/teacher'
+# require_relative 'classes/student'
 require_relative 'classes/rental'
+require_relative 'modules/person_creator'
+require_relative 'modules/book_creator'
 
 class App
+  include Creator
   def initialize
     @books = []
     @persons = []
@@ -16,65 +19,16 @@ class App
     system('clear')
   end
 
-  def create_teacher
-    puts '-creating a new teacher-'
-    print 'Name: '
-    name = gets.chomp
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'specialization: '
-    spec = gets.chomp
-    @persons.push(Teacher.new(age, spec, name))
-    puts "Person created successfully\n"
-  end
-
-  def permission?
-    print "Has parents' permission? [Y/N]: "
-    option = gets.chomp.capitalize
-    case option
-    when 'N'
-      false
-    when 'Y'
-      true
-    else
-      permission?
-    end
-  end
-
-  def create_student
-    puts '-creating a new student-'
-    print 'Name: '
-    name = gets.chomp
-    print 'Age: '
-    age = gets.chomp.to_i
-    permission = true
-    permission = permission? if age < 18
-    @persons.push(Student.new(age, name, parent_permission: permission))
-    puts "Person created successfully\n"
-  end
-
   def create_person
     clear_screen
-    print 'Do you want to create a student (1) or a teacher (2)? [Input the number]: '
-    option = gets.chomp.to_i
-    case option
-    when 1
-      create_student
-    when 2
-      create_teacher
-    else
-      create_person
-    end
+    @persons.push(create_new_person)
+    puts "Person created successfully\n"
   end
 
   def create_book
     clear_screen
-    puts '-creating new book-'
-    print 'Title: '
-    title = gets.chomp
-    print 'Author: '
-    author = gets.chomp
-    @books.push(Book.new(title, author))
+    @books.push(create_new_book)
+    puts "Book created successfully\n"
   end
 
   def select_book
@@ -89,7 +43,9 @@ class App
 
   def select_person
     puts 'Select a person from the list by [number]:'
-    @persons.each_with_index { |person, i| puts "[#{i}] [#{person.class.name}] ID: #{person.id} Name: #{person.name}" }
+    @persons.each_with_index do |person, i|
+      puts "[#{i}] [#{person.class.name}] ID: #{person.id} Name: #{person.name}"
+    end
     print 'Enter [number]: '
     index = gets.chomp.to_i
     return @persons[index] if @persons[index]
