+++ modules/book_creator.rb
@@ -1,0 +1,12 @@
+require_relative '../classes/book'
+
+module BookCreator
+  def create_libro
+    puts '-creating new book-'
+    print 'Title: '
+    title = gets.chomp
+    print 'Author: '
+    author = gets.chomp
+    Book.new(title, author)
+  end
+end
